Correct typos in the storage interface's phantom types
A connection is either read/write (type rw = [`R|`W]) or read/only
(type ro = [`R]). A function which requires the ability to write
needs to take a parameter [> `W]. In a few storage functions this
was typoed as [`W] which meant that a read/write connection couldn't
call the functions.

This affects:
  P.set_autostart
  V.create_xml
  V.delete

Signed-off-by: David Scott <dave.scott@eu.citrix.com>